Add merge function for merging classes together
lib.mkMerge is problematic because it doesn't actually merge the values
together. Instead, it creates a special attribute set that
lib.evalModules knows how to interpret. This means that if we have
something like

rec
{ c1.background = "blue";
  c2.color = "red";
  c3 = lib.mkMerge [ c1 c2 ];
  c4 = c3 // { font-size = "16px"; };
}

c1, c2, and c3 will behave as intended, but c4 will silently do
something different.

The merge function introduced in this commit will actually create a new
value that can be used like any other.